Allow for authorization in RPC Client
Adds a 'AuthHeader' property to the RPC Client struct. This is added to RPC requests as the Authorization header, and means private Nano nodes can be used. It's an option in the Nault wallet I found useful. I've not tested it here, but it shouldn't be a breaking change, and the Client should handle 401/403 errors as it normally would.